ajustes mais validação extra

+++ src/Model/Peca.php
@@ -157,13 +157,18 @@ class Peca
     {
         $con = Db::getConnection();
         $posto = intval($posto);
+        $peca = intval($peca);
 
-        $sql = "DELETE FROM tbl_peca WHERE peca = $peca AND posto = $posto";
-        $res = pg_query($con, $sql);
+        $peca_tem_estoque = self::pecaTemEstoque($peca, $posto);
 
-        return $res
-            ? ['status' => 'success', 'message' => 'Peça excluido com sucesso']
-            : ['status' => 'error', 'message' => 'Erro ao exlcuir peça.'];
+        if ($peca_tem_estoque == false) {
+            $sql = "DELETE FROM tbl_peca WHERE peca = $peca AND posto = $posto";
+            $res = pg_query($con, $sql);
+
+            return $res ? ['status' => 'success', 'message' => 'Peça excluída com sucesso.'] : ['status' => 'error', 'message' => 'Erro ao excluir peça.'];
+        } else {
+            return ['status' => 'error', 'message' => 'Não é possível excluir. A peça ainda está vinculada ao estoque.'];
+        }
     }
 
     public static function autocompletePecas($termo, $posto)
@@ -194,5 +199,26 @@ class Peca
             ];
         }
         return $sugestoes;
+    }
+
+    private static function pecaTemEstoque($peca, $posto)
+    {
+        $con = Db::getConnection();
+        $peca = intval($peca);
+        $posto = intval($posto);
+
+        $sql = "SELECT qtde
+                FROM tbl_estoque
+                WHERE peca = $peca
+                AND posto = $posto
+            ";
+        $res = pg_query($con, $sql);
+
+        if (pg_num_rows($res) > 0) {
+            $row = pg_fetch_assoc($res);
+            return $row['qtde'] > 0;
+        }
+
+        return false;
     }
 }
